+++ example/peripheral/media/media_test/src/cmd_media.c
@@ -31,7 +31,6 @@
 #include "shell.h"
 #include "shell_port.h"
 
-#include "fdc_common_hw.h"
 #include "media_example.h"